backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
gunicorn==22.0.0
python-dotenv==1.0.1

psycopg2-binary==2.9.9
redis==5.0.8
SQLAlchemy==2.0.35
alembic==1.13.3

anthropic==0.40.0
deepgram-sdk==3.7.0
openai==1.54.0

requests==2.32.3
httpx==0.27.2
pydantic==2.9.2

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

pillow==10.4.0
pydub==0.25.1
numpy==1.26.4

python-multipart==0.0.12
aiofiles==24.1.0

# Voice biometrics (Resemblyzer + pgvector)
resemblyzer==0.1.4
scipy==1.14.1
torch==2.4.1
pgvector==0.3.6

structlog==24.4.0
rich==13.9.2
